implement timedeltas.test_scalar_compat
